Change Query to View
There has been significant confusion around the use of Query and QueryProcessor and what the relation is with a 'View'.
The common opinion seems to be that 'View' expreses the concept more precisely and with that in mind we've made two breaking changes in naming:
- `Query` is now `View`
- `QueryProcessor` is now `Query`

My apologies in advance for the number of fixes that will be needed, we're still working through them here.